- sync.sh improvements for parallel execution
- MOVED_TO event updated to recognize and handle missing events
- some documentation, cleanup and spelling

+++ start.sh
@@ -5,7 +5,7 @@ if [[ $LC_ALL != "en_US.UTF-8" ]]; then
   locale-gen "${LC_ALL}"
 fi
 
-# Check mandatory veriables
+# Check mandatory variables and store them to secrets file
 if [ -z ${WEBDRIVE_USER} ]; then
   echo "[ERROR] Webdrive user is not set!"
   exit 1
@@ -67,27 +67,28 @@ trap "container_exit SIGTERM" SIGTERM
 trap "container_exit SIGINT" SIGINT
 
 
-echo "[INFO] Start completed. Start initital syncronization and filewatcher"
+echo "[INFO] Start completed. Start initial synchronization and file watcher"
 echo "===================================================================================================="
 
 
-# initial synchronization, perfomed in background
+# start the initial synchronization as background job
 # this script prints output in container logs, when finished
-nohup bash sync.sh "$SOURCE_DIR" "$WEBDRIVE_DIR" &
+# usage sync.sh: $1 = source | $2 = destination | $3 = reason
+/bin/bash sync.sh "$SOURCE_DIR" "$WEBDRIVE_DIR" "container-start" &
 
 
-# setting up filewatcher and actions for for high-performance instant synchronization per-event
-# supports renaming and file-move, to preserve existing files in nextcloud (instead of delete+recreate)
+# setting up file watcher and actions for for high-performance instant synchronization per-event
+# supports renaming and file-move, to preserve existing files in Nextcloud (instead of delete+recreate)
 inotifywait -m -r -e modify,create,delete,move "$SOURCE_DIR" --format '%e|%w%f|%f' |
 while IFS='|' read -r event full_path filename; do
   RELATIVE_PATH="${full_path/${SOURCE_DIR}\//''}"
   case "$event" in
     MODIFY|CREATE)
-      echo "[ACTION] Detected $event-Event - Copying: $filename"
+      echo "[ACTION] Detected $event-Event - Copying file: $filename"
       cp "$SOURCE_DIR/$RELATIVE_PATH" "$WEBDRIVE_DIR/$RELATIVE_PATH" --verbose
       ;;
     DELETE)
-      echo "[ACTION] Detected $event-Event - Deleting: $filename"
+      echo "[ACTION] Detected $event-Event - Deleting file: $filename"
       rm "$WEBDRIVE_DIR/$RELATIVE_PATH" --verbose
       ;;
     CREATE,ISDIR)
@@ -100,25 +101,33 @@ while IFS='|' read -r event full_path filename; do
       ;;
     MOVED_FROM)
       echo "[INFO] Detected $event-Event - File moved: $RELATIVE_PATH"
-      #OLD_PATH_LOCAL="$SOURCE_DIR/$RELATIVE_PATH"
       OLD_PATH_WEBDRIVE="$WEBDRIVE_DIR/$RELATIVE_PATH"
       ;;
     MOVED_TO)
-      echo "[ACTION] Detected $event-Event - File moved: $RELATIVE_PATH"
-      #NEW_PATH_LOCAL="$SOURCE_DIR/$RELATIVE_PATH"
+      echo "[ACTION] Detected $event-Event - Moving file: $RELATIVE_PATH"
       NEW_PATH_WEBDRIVE="$WEBDRIVE_DIR/$RELATIVE_PATH"
-      if [[ -n "$OLD_PATH_WEBDRIVE" ]]; then
-        mv "$OLD_PATH_WEBDRIVE" "$NEW_PATH_WEBDRIVE" --verbose
-        #NEW_PATH_LOCAL=""
-        NEW_PATH_WEBDRIVE=""
+      if [[ -n "$OLD_PATH_WEBDRIVE" && -f "$OLD_PATH_WEBDRIVE" ]]; then
+          mv "$OLD_PATH_WEBDRIVE" "$NEW_PATH_WEBDRIVE" --verbose
       else
-        echo "[ERROR] Variable \"OLD_PATH_WEBDRIVE\" not set"
+        if [[ ! -n "$OLD_PATH_WEBDRIVE" ]]; then
+          echo "[WARNING] Variable \"OLD_PATH_WEBDRIVE\" not set! Copying as new file!"
+        fi
+        if [[ ! -f "$OLD_PATH_WEBDRIVE" ]]; then
+          echo "[WARNING] File from MOVED_FROM event does not exist! Copying as new file!"
+        fi
+        cp "$SOURCE_DIR/$RELATIVE_PATH" "$WEBDRIVE_DIR/$RELATIVE_PATH" --verbose
+        echo "[INFO] Start complete sync run to fix other potential failures"
+        /bin/bash sync.sh "$SOURCE_DIR" "$WEBDRIVE_DIR" "missed-events" &
+        # usage sync.sh: $1 = source | $2 = destination | $3 = reason
       fi
+      unset OLD_PATH_WEBDRIVE
+      unset NEW_PATH_WEBDRIVE
       ;;
     *)
       echo "[ERROR] Unknown $event-Event for $filename"
       echo "Full path: $full_path"
       ;;
   esac
+  unset RELATIVE_PATH
 done &
 wait
